fix: address review comments for cache_control support
- Add cache_control to DEFAULT_CHAT_COMPLETION_PARAM_VALUES in constants.py
  so top-level litellm.completion(cache_control={...}) is not filtered out
- Consolidate tests into existing chat/test_anthropic_chat_transformation.py
- Remove duplicate test file (fixes check-duplicates CI failure)
- Remove unnecessary sys.path manipulation (fixes Ruff E402 lint failure)
